+++ src/transformer/type-literal.ts
@@ -1,6 +1,6 @@
 import * as ts from 'typescript';
 import { F_OPTIONAL, RtFunctionType, RtObjectMember, RtParameter, RtSerialized, RtType, T_ANY, T_ARRAY, T_ENUM, T_FALSE, T_FUNCTION, T_GENERIC, T_INTERSECTION, T_MAPPED, T_NULL,
-    T_OBJECT, T_THIS, T_TRUE, T_TUPLE, T_UNDEFINED, T_UNION, T_UNKNOWN, T_VOID } from '../common';
+    T_OBJECT, T_STAND_IN, T_THIS, T_TRUE, T_TUPLE, T_UNDEFINED, T_UNION, T_UNKNOWN, T_VOID } from '../common';
 import { findRelativePathToFile } from './find-relative-path';
 import { getPreferredExportForImport } from './get-exports-for-symbol';
 import { literalNode } from './literal-node';
@@ -103,7 +103,8 @@ export function typeLiteral(encoder: TypeEncoderImpl, type: ts.Type, typeNode?: 
                 return serialize({
                     TΦ: T_MAPPED,
                     t: literalNode(encoder.referToType(typeRef.typeParameter)),
-                    p: typeRef.aliasTypeArguments?.map(t => literalNode(encoder.referToType(t))) ?? []
+                    p: typeRef.aliasTypeArguments?.map(t => literalNode(encoder.referToType(t))) ?? [],
+                    m: serializeObjectMembers(typeRef, typeNode, encoder)
                 });
             }
         } else if ((objectType.objectFlags & ts.ObjectFlags.Reference) !== 0) {
@@ -284,21 +285,20 @@ export function typeLiteral(encoder: TypeEncoderImpl, type: ts.Type, typeNode?: 
 function serializeObjectMembers(type: ts.Type, typeNode: ts.TypeNode, encoder: TypeEncoderImpl) {
     let members: RtObjectMember[] = [];
 
-    if (type.symbol && type.symbol.members) {
-        type.symbol.members.forEach((value, key) => {
-            // TODO: currentTopStatement may be far up the AST- would be nice if we had
-            // a currentStatement that was not constrained to be at the top of the SourceFile
-            let memberType = encoder.ctx.checker.getTypeOfSymbolAtLocation(
-                value,
-                typeNode ?? encoder.ctx.currentTopStatement
-            );
+    let props = type.getProperties();
+    for (let prop of props) {
+        // TODO: currentTopStatement may be far up the AST- would be nice if we had
+        // a currentStatement that was not constrained to be at the top of the SourceFile
+        let memberType = encoder.ctx.checker.getTypeOfSymbolAtLocation(
+            prop,
+            typeNode ?? encoder.ctx.currentTopStatement
+        );
 
-            members.push({
-                n: <string>key,
-                f: `${hasFlag(value.flags, ts.SymbolFlags.Optional) ? F_OPTIONAL : ''}`,
-                t: <any>literalNode(encoder.referToType(memberType))
-            })
-        });
+        members.push({
+            n: prop.name,
+            f: `${hasFlag(prop.flags, ts.SymbolFlags.Optional) ? F_OPTIONAL : ''}`,
+            t: <any>literalNode(encoder.referToType(memberType))
+        })
     }
 
     return members;
@@ -343,6 +343,17 @@ export function referToTypeWithIdentifier(ctx: RttiContext, type: ts.Type, typeN
 
         if (typeHasValue(type)) {
             let entityName = checker.symbolToEntityName(type.symbol, ts.SymbolFlags.Class, undefined, undefined);
+            if (!entityName) {
+                /**
+                 * For instance, anonymous class expressions. When the class expression is defined at runtime its
+                 * type table slot will be replaced with the constructor reference, so we emit a T_STAND_IN.
+                 */
+                return serialize({
+                    TΦ: T_STAND_IN,
+                    name: undefined,
+                    note: `(anonymous class)`
+                });
+            }
             expr = serializeEntityNameAsExpression(entityName, containingSourceFile);
         } else {
             let symbolName = `IΦ${type.symbol.name}`;