+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # vNext
 - Typescript 4.7 support
+- Support inspecting properties of mapped types, and uses this metadata in matchesValue()
 
 # v0.8.2
 - Fix duplicate items in `properties`, `methods`, `staticProperties` and `staticMethods` (and the corresponding `*Names`